Move chrono-specific code to a separate header
