+++ CMakeLists.txt
@@ -138,8 +138,8 @@ function(add_headers VAR)
 endfunction()
 
 # Define the fmt library, its includes and the needed defines.
-add_headers(FMT_HEADERS color.h core.h format.h format-inl.h locale.h ostream.h
-                        printf.h time.h ranges.h)
+add_headers(FMT_HEADERS chrono.h color.h core.h format.h format-inl.h locale.h
+                        ostream.h printf.h time.h ranges.h)
 set(FMT_SOURCES src/format.cc)
 if (HAVE_OPEN)
   add_headers(FMT_HEADERS posix.h)
