+++ include/fmt/time.h
@@ -12,11 +12,6 @@
 #include <ctime>
 #include <locale>
 
-#if FMT_HAS_INCLUDE(<chrono>)
-# include <chrono>
-# include <sstream>
-#endif
-
 FMT_BEGIN_NAMESPACE
 
 // Prevents expansion of a preceding token as a function-style macro.
@@ -28,241 +23,7 @@ inline null<> localtime_r FMT_NOMACRO(...) { return null<>(); }
 inline null<> localtime_s(...) { return null<>(); }
 inline null<> gmtime_r(...) { return null<>(); }
 inline null<> gmtime_s(...) { return null<>(); }
-
-enum class numeric_system {
-  standard,
-  // Alternative numeric system, e.g. 十二 instead of 12 in ja_JP locale.
-  alternative
-};
-
-// Parses a put_time-like format string and invokes handler actions.
-template <typename Char, typename Handler>
-FMT_CONSTEXPR const Char *parse_chrono_format(
-    const Char *begin, const Char *end, Handler &&handler) {
-  auto ptr = begin;
-  while (ptr != end) {
-    auto c = *ptr;
-    if (c == '}') break;
-    if (c != '%') {
-      ++ptr;
-      continue;
-    }
-    if (begin != ptr)
-      handler.on_text(begin, ptr);
-    ++ptr; // consume '%'
-    if (ptr == end)
-      throw format_error("invalid format");
-    c = *ptr++;
-    switch (c) {
-    case '%':
-      handler.on_text(ptr - 1, ptr);
-      break;
-    // Day of the week:
-    case 'a':
-      handler.on_abbr_weekday();
-      break;
-    case 'A':
-      handler.on_full_weekday();
-      break;
-    case 'w':
-      handler.on_dec0_weekday(numeric_system::standard);
-      break;
-    case 'u':
-      handler.on_dec1_weekday(numeric_system::standard);
-      break;
-    // Month:
-    case 'b':
-      handler.on_abbr_month();
-      break;
-    case 'B':
-      handler.on_full_month();
-      break;
-    // Hour, minute, second:
-    case 'H':
-      handler.on_24_hour(numeric_system::standard);
-      break;
-    case 'I':
-      handler.on_12_hour(numeric_system::standard);
-      break;
-    case 'M':
-      handler.on_minute(numeric_system::standard);
-      break;
-    case 'S':
-      handler.on_second(numeric_system::standard);
-      break;
-    // Alternative numeric system:
-    case 'O':
-      if (ptr == end)
-        throw format_error("invalid format");
-      c = *ptr++;
-      switch (c) {
-      case 'w':
-        handler.on_dec0_weekday(numeric_system::alternative);
-        break;
-      case 'u':
-        handler.on_dec1_weekday(numeric_system::alternative);
-        break;
-      case 'H':
-        handler.on_24_hour(numeric_system::alternative);
-        break;
-      case 'I':
-        handler.on_12_hour(numeric_system::alternative);
-        break;
-      case 'M':
-        handler.on_minute(numeric_system::alternative);
-        break;
-      case 'S':
-        handler.on_second(numeric_system::alternative);
-        break;
-      }
-      break;
-      // TODO: parse more format specifiers
-    }
-    begin = ptr;
-  }
-  if (begin != ptr)
-    handler.on_text(begin, ptr);
-  return ptr;
-}
-
-struct chrono_format_checker {
-  template <typename Char>
-  void on_text(const Char *, const Char *) {}
-  void on_abbr_weekday() {}
-  void on_full_weekday() {}
-  void on_dec0_weekday(numeric_system) {}
-  void on_dec1_weekday(numeric_system) {}
-  void on_abbr_month() {}
-  void on_full_month() {}
-  void on_24_hour(numeric_system) {}
-  void on_12_hour(numeric_system) {}
-  void on_minute(numeric_system) {}
-  void on_second(numeric_system) {}
-};
 }  // namespace internal
-
-#ifdef __cpp_lib_chrono
-namespace internal {
-
-template <typename Int>
-inline int to_int(Int value) {
-  FMT_ASSERT(value >= (std::numeric_limits<int>::min)() &&
-             value <= (std::numeric_limits<int>::max)(), "invalid value");
-  return static_cast<int>(value);
-}
-
-template <typename FormatContext>
-struct chrono_formatter {
-  FormatContext &context;
-  typename FormatContext::iterator out;
-  std::chrono::seconds s;
-  std::chrono::milliseconds ms;
-
-  using char_type = typename FormatContext::char_type;
-
-  explicit chrono_formatter(FormatContext &ctx)
-    : context(ctx), out(ctx.out()) {}
-
-  void write(int value, int width) {
-    typedef typename int_traits<int>::main_type main_type;
-    main_type n = to_unsigned(value);
-    int num_digits = static_cast<int>(internal::count_digits(n));
-    if (width > num_digits)
-      out = std::fill_n(out, width - num_digits, '0');
-    out = format_decimal<char_type>(out, n, num_digits);
-  }
-
-  void format_localized(const tm &time, char format) {
-    auto locale = context.locale().template get<std::locale>();
-    auto &facet = std::use_facet<std::time_put<char_type>>(locale);
-    std::basic_ostringstream<char_type> os;
-    os.imbue(locale);
-    const char format_str[] = {'%', 'O', format};
-    facet.put(os, os, ' ', &time, format_str, format_str + sizeof(format_str));
-    auto str = os.str();
-    std::copy(str.begin(), str.end(), out);
-  }
-
-  void on_text(const char_type *begin, const char_type *end) {
-    std::copy(begin, end, out);
-  }
-
-  void on_abbr_weekday() {}
-  void on_full_weekday() {}
-  void on_dec0_weekday(numeric_system) {}
-  void on_dec1_weekday(numeric_system) {}
-  void on_abbr_month() {}
-  void on_full_month() {}
-
-  void on_24_hour(numeric_system ns) {
-    auto hour = to_int((s.count() / 3600) % 24);
-    if (ns == numeric_system::standard)
-      return write(hour, 2);
-    auto time = tm();
-    time.tm_hour = hour;
-    format_localized(time, 'H');
-  }
-
-  void on_12_hour(numeric_system ns) {
-    auto hour = to_int((s.count() / 3600) % 12);
-    hour = hour > 0 ? hour : 12;
-    if (ns == numeric_system::standard)
-      return write(hour, 2);
-    auto time = tm();
-    time.tm_hour = hour;
-    format_localized(time, 'I');
-  }
-
-  void on_minute(numeric_system ns) {
-    auto minute = to_int((s.count() / 60) % 60);
-    if (ns == numeric_system::standard)
-      return write(minute, 2);
-    auto time = tm();
-    time.tm_min = minute;
-    format_localized(time, 'M');
-  }
-
-  void on_second(numeric_system ns) {
-    auto second = to_int(s.count() % 60);
-    if (ns == numeric_system::standard) {
-      write(second, 2);
-      if (ms != std::chrono::milliseconds()) {
-        *out++ = '.';
-        write(to_int(ms.count()), 3);
-      }
-      return;
-    }
-    auto time = tm();
-    time.tm_sec = second;
-    format_localized(time, 'S');
-  }
-};
-}  // namespace internal
-
-template <typename Rep, typename Period, typename Char>
-struct formatter<std::chrono::duration<Rep, Period>, Char> {
-  mutable basic_string_view<Char> format_str;
-  using Duration = std::chrono::duration<Rep, Period>;
-
-  FMT_CONSTEXPR auto parse(basic_parse_context<Char> &ctx)
-      -> decltype(ctx.begin()) {
-    auto begin = ctx.begin(), end = ctx.end();
-    end = parse_chrono_format(begin, end, internal::chrono_format_checker());
-    format_str = basic_string_view<Char>(&*begin, end - begin);
-    return end;
-  }
-
-  template <typename FormatContext>
-  auto format(const Duration &d, FormatContext &ctx)
-      -> decltype(ctx.out()) {
-    internal::chrono_formatter<FormatContext> f(ctx);
-    f.s = std::chrono::duration_cast<std::chrono::seconds>(d);
-    f.ms = std::chrono::duration_cast<std::chrono::milliseconds>(d - f.s);
-    parse_chrono_format(format_str.begin(), format_str.end(), f);
-    return f.out;
-  }
-};
-#endif  // __cpp_lib_chrono
 
 // Thread-safe replacement for std::localtime
 inline std::tm localtime(std::time_t time) {
